Intermediate Code Push
