+++ TelCo.ColorCoder/Colors.cs
@@ -1,0 +1,44 @@
+﻿using System.Drawing;
+
+namespace TelCo.ColorCoder
+{
+
+    /// <summary>
+    /// data type defined to hold the two colors of clor pair
+    /// </summary>
+    internal class ColorPair
+    {
+        internal Color majorColor;
+        internal Color minorColor;
+        public override string ToString()
+        {
+            return string.Format("MajorColor:{0}, MinorColor:{1}", majorColor.Name, minorColor.Name);
+        }
+    }
+
+    internal static class ColorMaps
+    {
+        /// <summary>
+        /// Array of Major colors
+        /// </summary>
+        public static Color[] colorMapMajor { get; set; }
+        /// <summary>
+        /// Array of minor colors
+        /// </summary>
+        public static Color[] colorMapMinor { get; set; }
+
+  
+        /// <summary>
+        /// Static constructor required to initialize static variable
+        /// </summary>
+
+        static ColorMaps()
+        {
+            colorMapMajor = new Color[] { Color.White, Color.Red, Color.Black, Color.Yellow, Color.Violet };
+            colorMapMinor = new Color[] { Color.Blue, Color.Orange, Color.Green, Color.Brown, Color.SlateGray };
+        }
+
+    }
+}
+
+
